Add Travian::Attack class first sketch
Implements the attack class which contains a helper method to parse
attack times of the format hh:mm:ss.
Based on the time parsed it allows access to that time object through
Attack#at and the number of seconds until the attack via Attack#in.
Also add Attack.incoming? which returns true if there are any incoming
attacks

Signed-off-by: João Soares <jsoaresgeral@gmail.com>