Add Travian::Resource base class
Resource is the base class to describe a resource amount composed by
all 4 types(wood, clay, iron, cereal).
It defines the unary minus operator for resources, the sum and
subtraction with another resource and the multiplication operator to
multiply by an integer.
It includes the Enumerable module and defines the required each method.
It includes the Comparable module and defines the required spaceship
operator.
It also redefines the #eql? method to the same meaning of #== and also
the hash method conforming to the above.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>